GaleriActivity eklendi. Ve galeriden resim yükleme özelliği eklendi.

+++ imageeditorlib/src/main/java/com/passiondroid/imageeditorlib/ImageEditActivity.java
@@ -1,5 +1,6 @@
 package com.passiondroid.imageeditorlib;
 
+import static com.passiondroid.imageeditorlib.ImageEditor.EXTRA_CACHE;
 import static com.passiondroid.imageeditorlib.ImageEditor.EXTRA_IMAGE_PATH;
 
 import android.app.Activity;
@@ -17,6 +18,7 @@ import java.io.File;
 public class ImageEditActivity extends BaseImageEditActivity
     implements PhotoEditorFragment.OnFragmentInteractionListener {
   private String oldImagePath;
+  private boolean isCache = true;
   //private View touchView;
 
   @Override
@@ -25,6 +27,7 @@ public class ImageEditActivity extends BaseImageEditActivity
     setContentView(R.layout.activity_image_edit);
 
     oldImagePath = getIntent().getStringExtra(EXTRA_IMAGE_PATH);
+    isCache = getIntent().getBooleanExtra(EXTRA_CACHE, true);
     if (oldImagePath != null) {
       FragmentUtil.addFragment(this, R.id.fragment_container,
           PhotoEditorFragment.newInstance(oldImagePath));
@@ -59,9 +62,11 @@ public class ImageEditActivity extends BaseImageEditActivity
   }
 
   private void DeletePhoto() {
-    File oldImage = new File(oldImagePath);
-    if (oldImage.exists()){
-      boolean b = oldImage.delete();
+    if (isCache){
+      File oldImage = new File(oldImagePath);
+      if (oldImage.exists()){
+        boolean b = oldImage.delete();
+      }
     }
   }
 }
